requirements.txt
fastapi
uvicorn[standard]
pymupdf
pdf2image
pytesseract
pillow
pandas